fixing qa
