+++ messages/uk.json
@@ -105,6 +105,8 @@
     },
     "Q&A": {
         "pageTitle": "q&a",
+        "mobileBigTitlePart1": "Відповіді на ",
+        "mobileBigTitlePart2": "часті питання",
         "pageBigTitle": "Відповіді на часті питання",
         "question1": "Що обрати - лендінг чи повноцінний сайт?",
         "answer1": "На нашому сайті ви знайдете інформацію про вже реалізовані проекти та команди, які стояли за їхнім створенням, а також детальні профілі кожного учасника команди з описом їхніх навичок, технологічного стеку, посиланнями на соціальні мережі та контактними даними. ",
