+++ src/components/qAsection/Accordion.tsx
@@ -21,7 +21,7 @@ export function Accordion({ item }: AccordionProps) {
             <button onClick={changeStateAccordion} className="w-full">
                 <div className="flex justify-between my-[16px] pc:mb-[32px]">
                     <div>
-                        <summary className="flex-none font-caviar self-center h-auto text-left py-0 text-baseb pc:text-xlb ">
+                        <summary className="flex-none font-caviar text-white self-center h-auto text-left py-0 text-baseb pc:text-xlb ">
                             {item.question}
                         </summary>
                     </div>
@@ -38,7 +38,7 @@ export function Accordion({ item }: AccordionProps) {
             </button>
 
             <p
-                className={`text-geist text-sm overflow-hidden mb-[8px] transition-[max-height] duration-[300ms] ease-in-out 
+                className={`text-geist text-sm tab:text-base deskxl:text-lg overflow-hidden mb-[8px] transition-[max-height] duration-[300ms] ease-in-out 
                     ${isAccordionOpen ? "max-h-[300px]" : "max-h-0"}`}
             >
                 {item.answer}
