+++ src/components/qAsection/QaSection.tsx
@@ -1,5 +1,5 @@
 import { Section } from "../shared/Section";
-import { TitleQA } from "./TitleQA";
+import { TitleQA } from "./TitleQa/TitleQA";
 import { QuestionsList } from "./QuestionsList";
 import { SharePopover } from "../shared/SharePopover";
 
@@ -9,7 +9,7 @@ export const QaSection = () => {
             <div className="flex">
                 <TitleQA />
                 <SharePopover
-                    className="absolute -top-3 right-4 tab:top-0 tab:right-4 pc:right-[60px]"
+                    className="absolute -top-3 right-4 tab:top-0 tab:right-4 pc:right-[60px] pc:top-[25px] deskxl:top-[76px]"
                     trigerShowShareText={true}
                 />
             </div>
